Edit readme and transfer QueueItem class into the same file as inverted_index file
